Move client-interface into client
