+++ Cargo.toml
@@ -10,7 +10,6 @@ edition = "2018"
 members = [
   "cli",
   "client",
-  "client-interface",
   "node",
   "runtime",
   "subxt",
