+++ DEVELOPING.md
@@ -24,8 +24,6 @@ Packages
 * `node` contains the node code which includes the runtime code.
 * `client` contains the high-level client library for interacting with the
   registry through a node and a memory test double.
-* `client-interface` defines the trait for clients and provides the necessary
-  data types.
 * `cli` contains a binary for interacting with the registry node to submit
   transactions and read state.
 * `subxt` contains a copy of [`subxt`][subxt], the Rust client library for
