--- client-interface/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "radicle_registry_client_interface"
-version = "0.1.0"
-authors = ["Thomas Scholtes <thomas@monadic.xyz>"]
-edition = "2018"
-
-[dependencies]
-radicle_registry_runtime = { path = "../runtime" }
-substrate-subxt = { path = "../subxt" }
-
-parity-scale-codec = "1.0"
-futures = "0.1"
-
-[dependencies.paint-system]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.paint-transaction-payment]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.substrate-primitives]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.sr-primitives]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.paint-support]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
-
-[dependencies.sr-io]
-git = "https://github.com/paritytech/substrate"
-rev = "dc4a8329b9c47790c25429a5df272b3648275c1b"
